fix: 중ã충돌 해결
