+++ backend/src/main/java/com/team3/airdnd/accommodation/controller/AccommodationController.java
@@ -1,12 +1,14 @@
 package com.team3.airdnd.accommodation.controller;
 
 import com.team3.airdnd.accommodation.dto.AccommodationResponseDto;
-import com.team3.airdnd.accommodation.service.AccommodationService;
 import com.team3.airdnd.global.dto.ResponseDto;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -20,5 +22,14 @@ public class AccommodationController {
     public ResponseDto<AccommodationResponseDto.AccommodationDetailDto> getAccommodationDetail(@PathVariable("accommodation-id") Long id) {
         AccommodationResponseDto.AccommodationDetailDto detailDto = accommodationService.getAccommodationDetail(id);
         return ResponseDto.ok(detailDto);
-    }
+	}
+
+	@GetMapping("")
+	public ResponseDto<AccommodationResponseDto.AccommodationListDto> getAccommodationList(
+		@RequestParam(required = false, defaultValue = "1") int page,
+		@RequestParam(required = false, defaultValue = "5") int size
+        ){
+			AccommodationResponseDto.AccommodationListDto accommodations = accomodationService.getAccommodations(page, size);
+			return ResponseDto.ok(accommodations);
+		}
 }
