+++ backend/src/main/java/com/team3/airdnd/accommodation/dto/AccommodationResponseDto.java
@@ -44,4 +44,5 @@ public class AccommodationResponseDto {
         private int reviewSize;
         private List<ReviewInfoDto> comments;
     }
+
 }
