+++ backend/src/main/java/com/team3/airdnd/accommodation/repository/AccommodationAmenityRepository.java
@@ -21,6 +21,7 @@ public interface AccommodationAmenityRepository extends JpaRepository<Accommodat
         WHERE aa.accommodation.id = :accommodationId
   """)
     List<AmenityInfoDto> findAmenityByAccommodationId(
-        @Param("accommodationId") Long accommodationId
+            @Param("accommodationId") Long accommodationId
     );
+	List<AccommodationAmenity> findByAccommodation_Id(Long accommodationId);
 }
