+++ backend/src/main/java/com/team3/airdnd/accommodation/repository/AccommodationRepository.java
@@ -2,6 +2,7 @@ package com.team3.airdnd.accommodation.repository;
 
 import com.team3.airdnd.accommodation.domain.Accommodation;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
@@ -13,4 +14,5 @@ public interface AccommodationRepository extends JpaRepository<Accommodation, Lo
         WHERE a.id = :id
     """)
     Optional<Accommodation> findDetailById(@Param("id") Long id);
+
 }
