--- backend/src/main/java/com/team3/airdnd/accomodation/repository/AccommodationAmenityRepository.java
@@ -1,12 +1,0 @@
-package com.team3.airdnd.accomodation.repository;
-
-import java.util.List;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.team3.airdnd.accomodation.domain.AccommodationAmenity;
-
-
-public interface AccommodationAmenityRepository extends JpaRepository<AccommodationAmenity, Long> {
-	List<AccommodationAmenity> findByAccommodation_Id(Long accommodationId);;
-}
